requirements.txt
asgiref==3.7.2
astroid==3.3.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
Django==4.2.14
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pylint==3.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.13.2
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
